% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helicopter.R
\docType{data}
\name{helicopter}
\alias{helicopter}
\title{Toy dataset comparing Helicopter target identifications.}
\format{
A data frame with 20 rows and 2 variables:
\describe{
\item{apache}{Number of targets detected by the apache helicopter}
\item{viper}{Number of targets detected by the viper helicopter}
...
}
}
\source{
internally created toy dataset
}
\usage{
helicopter
}
\description{
A dataset containing the number of targets detected according to helicopter type.
}
\keyword{datasets}
